requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alpha-vantage==2.3.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
fonttools==4.39.4
frozendict==2.3.8
frozenlist==1.3.3
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pandas-datareader==0.10.0
Pillow==9.5.0
plotly==5.14.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tenacity==8.2.2
turingquant==0.2.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
yarl==1.9.2
yfinance==0.2.18
